functions/private/discord_bot/requirements.txt
discord-interactions
google-cloud-firestore